March 07, 2014 @ 2:47pm
Created an enemy spawner that does not go based on entering a trigger,
but rather goes based on how close the player is to the spawner. It is
called "Constant Spawner" and has a few more capabilities than the
"Trigger 'Spawner'". I placed a few sample ones around my scene to show
how they work. No mystery numbers inside the code - everything is
controlled by the inspector.