Sounds in the game
Sound Script and Sound Manager are complete. They need to be moved into
"The Truth" today. Therefore, I am uploading this to GitHub so that I
can add it to the truth in the other room. (This room is full of lies.)
The alternating music functionality is also going in. These are most
likely not the final sounds and only one of the songs is final. Deven
also needs to add the Ninja Script and the Sword Monk Script in order
for the enemies to make sounds when they die.